+++ providers/dns/stackpath/stackpath.go
@@ -6,19 +6,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
-	"net/url"
 	"time"
 
 	"github.com/go-acme/lego/v4/challenge/dns01"
 	"github.com/go-acme/lego/v4/log"
 	"github.com/go-acme/lego/v4/platform/config/env"
-	"golang.org/x/oauth2/clientcredentials"
-)
-
-const (
-	defaultBaseURL = "https://gateway.stackpath.com/dns/v1/stacks/"
-	defaultAuthURL = "https://gateway.stackpath.com/identity/v1/oauth2/token"
+	"github.com/go-acme/lego/v4/providers/dns/stackpath/internal"
 )
 
 // Environment variables names.
@@ -55,9 +48,8 @@ func NewDefaultConfig() *Config {
 
 // DNSProvider implements the challenge.Provider interface.
 type DNSProvider struct {
-	BaseURL *url.URL
-	client  *http.Client
-	config  *Config
+	config *Config
+	client *internal.Client
 }
 
 // NewDNSProvider returns a DNSProvider instance configured for Stackpath.
@@ -91,30 +83,18 @@ func NewDNSProviderConfig(config *Config) (*DNSProvider, error) {
 		return nil, errors.New("stackpath: stack id missing")
 	}
 
-	baseURL, _ := url.Parse(defaultBaseURL)
+	client := internal.NewClient(context.Background(), config.StackID, config.ClientID, config.ClientSecret)
 
-	return &DNSProvider{
-		BaseURL: baseURL,
-		client:  getOathClient(config),
-		config:  config,
-	}, nil
-}
-
-func getOathClient(config *Config) *http.Client {
-	oathConfig := &clientcredentials.Config{
-		TokenURL:     defaultAuthURL,
-		ClientID:     config.ClientID,
-		ClientSecret: config.ClientSecret,
-	}
-
-	return oathConfig.Client(context.Background())
+	return &DNSProvider{config: config, client: client}, nil
 }
 
 // Present creates a TXT record to fulfill the dns-01 challenge.
 func (d *DNSProvider) Present(domain, token, keyAuth string) error {
 	info := dns01.GetChallengeInfo(domain, keyAuth)
 
-	zone, err := d.getZones(info.EffectiveFQDN)
+	ctx := context.Background()
+
+	zone, err := d.client.GetZones(ctx, info.EffectiveFQDN)
 	if err != nil {
 		return fmt.Errorf("stackpath: %w", err)
 	}
@@ -124,21 +104,23 @@ func (d *DNSProvider) Present(domain, token, keyAuth string) error {
 		return fmt.Errorf("stackpath: %w", err)
 	}
 
-	record := Record{
+	record := internal.Record{
 		Name: subDomain,
 		Type: "TXT",
 		TTL:  d.config.TTL,
 		Data: info.Value,
 	}
 
-	return d.createZoneRecord(zone, record)
+	return d.client.CreateZoneRecord(ctx, zone, record)
 }
 
 // CleanUp removes the TXT record matching the specified parameters.
 func (d *DNSProvider) CleanUp(domain, token, keyAuth string) error {
 	info := dns01.GetChallengeInfo(domain, keyAuth)
 
-	zone, err := d.getZones(info.EffectiveFQDN)
+	ctx := context.Background()
+
+	zone, err := d.client.GetZones(ctx, info.EffectiveFQDN)
 	if err != nil {
 		return fmt.Errorf("stackpath: %w", err)
 	}
@@ -148,13 +130,13 @@ func (d *DNSProvider) CleanUp(domain, token, keyAuth string) error {
 		return fmt.Errorf("stackpath: %w", err)
 	}
 
-	records, err := d.getZoneRecords(subDomain, zone)
+	records, err := d.client.GetZoneRecords(ctx, subDomain, zone)
 	if err != nil {
 		return err
 	}
 
 	for _, record := range records {
-		err = d.deleteZoneRecord(zone, record)
+		err = d.client.DeleteZoneRecord(ctx, zone, record)
 		if err != nil {
 			log.Printf("stackpath: failed to delete TXT record: %v", err)
 		}
