+++ providers/dns/stackpath/stackpath_test.go
@@ -1,9 +1,6 @@
 package stackpath
 
 import (
-	"net/http"
-	"net/http/httptest"
-	"net/url"
 	"testing"
 	"time"
 
@@ -133,132 +130,6 @@ func TestNewDNSProviderConfig(t *testing.T) {
 			assert.Nil(t, p)
 		})
 	}
-}
-
-func setupTest(t *testing.T) (*DNSProvider, *http.ServeMux) {
-	t.Helper()
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	config := NewDefaultConfig()
-	config.ClientID = "CLIENT_ID"
-	config.ClientSecret = "CLIENT_SECRET"
-	config.StackID = "STACK_ID"
-
-	provider, err := NewDNSProviderConfig(config)
-	require.NoError(t, err)
-
-	provider.client = http.DefaultClient
-	provider.BaseURL, _ = url.Parse(server.URL + "/")
-
-	return provider, mux
-}
-
-func TestDNSProvider_getZoneRecords(t *testing.T) {
-	provider, mux := setupTest(t)
-
-	mux.HandleFunc("/STACK_ID/zones/A/records", func(w http.ResponseWriter, _ *http.Request) {
-		content := `
-			{
-				"records": [
-					{"id":"1","name":"foo1","type":"TXT","ttl":120,"data":"txtTXTtxt"},
-					{"id":"2","name":"foo2","type":"TXT","ttl":121,"data":"TXTtxtTXT"}
-				]
-			}`
-
-		_, err := w.Write([]byte(content))
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	})
-
-	records, err := provider.getZoneRecords("foo1", &Zone{ID: "A", Domain: "test"})
-	require.NoError(t, err)
-
-	expected := []Record{
-		{ID: "1", Name: "foo1", Type: "TXT", TTL: 120, Data: "txtTXTtxt"},
-		{ID: "2", Name: "foo2", Type: "TXT", TTL: 121, Data: "TXTtxtTXT"},
-	}
-
-	assert.Equal(t, expected, records)
-}
-
-func TestDNSProvider_getZoneRecords_apiError(t *testing.T) {
-	provider, mux := setupTest(t)
-
-	mux.HandleFunc("/STACK_ID/zones/A/records", func(w http.ResponseWriter, _ *http.Request) {
-		content := `
-{
-	"code": 401,
-	"error": "an unauthorized request is attempted."
-}`
-
-		w.WriteHeader(http.StatusUnauthorized)
-		_, err := w.Write([]byte(content))
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	})
-
-	_, err := provider.getZoneRecords("foo1", &Zone{ID: "A", Domain: "test"})
-
-	expected := &ErrorResponse{Code: 401, Message: "an unauthorized request is attempted."}
-	assert.Equal(t, expected, err)
-}
-
-func TestDNSProvider_getZones(t *testing.T) {
-	provider, mux := setupTest(t)
-
-	mux.HandleFunc("/STACK_ID/zones", func(w http.ResponseWriter, _ *http.Request) {
-		content := `
-{
-  "pageInfo": {
-    "totalCount": "5",
-    "hasPreviousPage": false,
-    "hasNextPage": false,
-    "startCursor": "1",
-    "endCursor": "1"
-  },
-  "zones": [
-    {
-      "stackId": "my_stack",
-      "accountId": "my_account",
-      "id": "A",
-      "domain": "foo.com",
-      "version": "1",
-      "labels": {
-        "property1": "val1",
-        "property2": "val2"
-      },
-      "created": "2018-10-07T02:31:49Z",
-      "updated": "2018-10-07T02:31:49Z",
-      "nameservers": [
-        "1.1.1.1"
-      ],
-      "verified": "2018-10-07T02:31:49Z",
-      "status": "ACTIVE",
-      "disabled": false
-    }
-  ]
-}`
-
-		_, err := w.Write([]byte(content))
-		if err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		}
-	})
-
-	zone, err := provider.getZones("sub.foo.com")
-	require.NoError(t, err)
-
-	expected := &Zone{ID: "A", Domain: "foo.com"}
-
-	assert.Equal(t, expected, zone)
 }
 
 func TestLivePresent(t *testing.T) {
